+++ src/main/java/controller/UserController.java
@@ -1,4 +1,4 @@
-package webserver;
+package controller;
 
 import db.Database;
 import http.*;
