+++ src/main/java/controller/ViewController.java
@@ -1,11 +1,10 @@
-package webserver;
+package controller;
 
 import http.*;
 
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
-import java.util.Arrays;
 
 public class ViewController implements Controller {
